split up autocompletion test
