+++ packages/language-support/src/autocompletion/completionCoreCompletions.ts
